[UNTESTED] User registration
THIS CHANGE IS NOT FULLY TESTED. Do NOT base changes off of this.

THIS CHANGE NEEDS REFACTORING. I finally figured out how to work with
the external JWT library, and thus there are more things that need
to be made before we can work cleanly with this.

THIS BRANCH IS NOT SUPPOSED TO BE HERE. I'm breaking the `dev` and
`master`-only rule because I need some cross-testing and validation
from @rishihot55 and others. DO NOT work on this branch or merge it
to any of the other branches. This branch will be deleted after work
and a PR-merge.

FUNCTIONALITY IS INCOMPLETE. Some more validation may need to be done,
and whole sets of fields are missing. If you have questions, ask
me on telegram.me/Mindstormer619


Signed-off-by: Siddarth Iyer <mindstormer619@gmail.com>