Add admin flag for users table and one seed user
Current implementation uses an admin flag. This may change in the future
to a full admin table, in which case we'll change the column to a
integer datatype instead of a boolean.

You can now run `php artisan migrate:refresh --seed` to obtain a new
user entry in the db.

~~ Siddarth