Add polymorphic model link between posts and events
Events can be "attached" (embedded) to posts, along with other future
entities. Hence wrt forward planning, it's a polymorphic relationship
which is illustrated in the model.

This completes the model relationships for all existing database models.
Votes, comments and circles are left.

PLEASE STICK TO THE FUNCTIONS DEFINED IN THE MODELS.
DO NOT USE RAW DB QUERIES OR WHERE CLAUSES UNLESS NECESSARY.

Signed-off-by: Siddarth Iyer <mindstormer619@gmail.com>